+++ files/HikerTest.kt
@@ -25,8 +25,8 @@ class HikerTest /*[X]*/ : StringSpec() {
         }
         
         "average value" {
-            hiker.average(arrayOf(6, 9, 15, -2, 92, 11)) shouldBe 21.833333.toFloat()
-            hiker.average(arrayOf(3, 3, 3, 3, 3, 3)) shouldBe 3.toFloat()
+            hiker.average(arrayOf(6, 9, 15, -2, 92, 11)) shouldBe 21.833333.toDouble()
+            hiker.average(arrayOf(3, 3, 3, 3, 3, 3)) shouldBe 3.toDouble()
         }
     }
 }
